Direct users to the /endpoints/ route immediately
Since we never came up with viable content for the "civic lab" identity/brand, let's just redirect users to the list of endpoints.
Someday we'll consolidate the civic-lab-service and endpoint-service content to a single API directory that points to the various API containers; until then, this is a good first step.