Remove redundant values, improve comments, delete redundant files
Removing redundant Default values in service.yaml files, where the value is always the same throughout the service.yaml.  As well and especially, to enable us to update the FQDN of the API containers without a ton of hassle or confusion.

Removing redundant (orphaned) Docker container /src/ files that were long ago migrated to standalone repos for each of the API containers.